+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Add query refId in response
 - Upgrade grafana-plugin-sdk-go to version v0.139.0.
 - Upgrade gosnowflake to version v1.6.13.
+- Upgrade js dependencies
 
 ## 1.1.0